- lottie lib 추가 - file provider 추가 - flower fragment 추가 - 일부 코드 정리 - context!! 를 requiredContext 로 변경 - onCommandEvent 의 인자 Any? 를 Any 로 변경

+++ common/src/main/java/com/example/common/ViewGroup.kt
@@ -53,7 +53,7 @@ inline fun TextView.gravityCenter() {
 inline fun View.dpToPx(v: Float) = v * context.displayDensity()
 inline fun View.pxToDp(v: Float) = v / context.displayDensity()
 inline fun View.spToPx(v: Float) = TypedValue.applyDimension(
-    TypedValue.COMPLEX_UNIT_SP, v.toFloat(), context.resources.displayMetrics)
+    TypedValue.COMPLEX_UNIT_SP, v, context.resources.displayMetrics)
 
 inline fun View.dpToPx(v: Int) = (v * context.displayDensity()).toInt()
 inline fun View.pxToDp(v: Int) = (v / context.displayDensity()).toInt()
